Replaced the PauseResumeListener interface with the BattleManager so it uses the IBattleListener thru the BattleEventDispatcher instead for onBattlePaused() and onBatteResumed().
Fixed some issues with the BattleView, which would flicker when starting a new battle, as the timer was not stop when the BattleObserver is disposed or garbage collected.

TODO: The "Next Step" button does not work, and must be fixed.

This is the first step of getting rid of GUI specific stuff in the Battle class, which does not really belongs to the battle.