Redesigned the RobocodeFrame.updateTitle() so it does not need access to the Battle instance. Now it uses only the BattleObserver. In addition, the timer was removed. The title is only updated every half second like before for battle turns.
In order to use the BattleSpecification provided with IBattleListener.onBattleStarted(), the BattleSpecification should never be null. So I had to create a change the BattleManager.startNewBattle(BattleProperties, boolean, boolean) to not use null, but instead call the new method BattleProperties.toBattleSpecification(). Note however, that the RobotSpecification[] is set to null on the returned BattleSpecification, as it is currently not possible to construct these. Currently this in no problem, as they are not used in this situation, i.e. our purpose.

The TurnSnapshot has been updated with getRound().