Bug [2828072] - Accel/decel rules introduced in 1.7.1.3 causes trouble. This bug has been undergoing a lots of discussions, alphas, and tests (going on for months). More detailed information can be found on these pages:
http://robowiki.net/w/index.php?title=Talk:Robocode/Game_Physics
http://robowiki.net/wiki/User:Positive/Optimal_Velocity
http://robowiki.net/wiki/User_talk:Voidious/Robocode_Version_Tests
http://robowiki.net/w/index.php?title=Talk:Darkcanuck/VelocityTest

This fix is a compromise of getting rid of old accel/decel rules, making the rules simpler and more correct, and in the same time stay as compatible with old RoboRumble robots as possible. The unit tests have been updated accordingly to reflect the changes in the robot movement, and hence scores. More unit tests will be added later, which are provided in the pages above.