Bug fix: .NET plugin unexpectedly accesses SMB and FTP when searching for robots
When Robocode launches, .NET plugin searches for robots by opening
a connection to a URI like file://C:/robots/robot.dll!/hoge.fuga that is
considered as a UNC path with the hostname "C".  Because Windows tries
a couple of protocols to connect to it, Robocode startup can be extremely
delayed, especially when the hostname "C" is found but invalid for SMB/FTP.

The fix is to add a prefix "file:///" instead of "file://".  This ensures
that the URI has an empty hostname and it's not considered as UNC.